Add files via upload
In our first year, we developed a Python project named "House Price Prediction." By using Linear Regression, we analyzed a Kaggle dataset to understand the relationships between various house attributes and prices. Our application asks users for input such as area, number of stories, and bedrooms desired, storing this data in a database. With this information, it predicts house prices.